@@ -32,43 +32,43 @@ dependencies:
     sdk: flutter
 
   # Supabase
-  supabase_flutter: ^2.5.0
+  supabase_flutter: ^2.12.0
 
   # Environment variables
   envied: ^0.5.4+1
 
   # State management
-  provider: ^6.1.2
+  provider: ^6.1.5
 
   # Utils
   flutter_hooks: ^0.20.5
-  package_info_plus: ^8.0.0
-  url_launcher: ^6.2.6
+  package_info_plus: ^8.3.0
+  url_launcher: ^6.3.0
   
-  auto_route: ^8.1.3
+  auto_route: ^8.1.4
   fpdart: ^1.1.0
-  dio: ^5.4.3+1
+  dio: ^5.7.0
 
   # 简繁中文转换（纯 Dart，支持 Web）
   pinyin: ^3.3.0
-  shared_preferences: ^2.2.3
+  shared_preferences: ^2.5.0
 
   # UI
   cached_network_image: ^3.4.1
-  google_fonts: ^6.2.1
+  google_fonts: ^6.3.3
   auto_size_text: ^3.0.0
-  jovial_svg: ^1.1.21-rc.2
-  flutter_markdown: ^0.7.1
+  jovial_svg: ^1.1.29
+  flutter_markdown: ^0.7.4
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^3.0.0
+  flutter_lints: ^5.0.0
   envied_generator: ^0.5.4+1
-  build_runner: ^2.4.9
+  build_runner: ^2.4.13
   auto_route_generator: ^8.0.0
-  flutter_launcher_icons: ^0.13.1
+  flutter_launcher_icons: ^0.14.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
